requirements.txt
langchain==0.1.0
langchain_community==0.0.10
faiss-cpu==1.7.4
torch==2.1.0
python-dotenv==1.0.0
colorama==0.4.6
transformers==4.34.0
tokenizers>=0.14.0,<0.15
huggingface_hub>=0.16.4,<1.0
sentence-transformers==2.2.2
